fix: resolve all audit findings from mail module review
Frontend:
- Mail Inbox: DOMPurify on all dangerouslySetInnerHTML, Full thread button in sent pane, RED constant for error color
- ThreadDetail: replace #10B981 with GREEN constant
- Compose: persist draft ID after first save; fix signature field name
- ReportBuilder, ScheduledReports: replace 'DM Mono' string with MONO constant
- EmailBlockEditor: replace 'DM Sans' with 'Inter' in email FONT constant
- Login: replace 'DM Mono' with 'Roboto Mono' in CSS
- App.tsx: import apiPost, use VITE_DEV_ROLE env var for dev mock role, CANVAS/RED for spinner, remove local MGMT const
- Sidebar: remove local MGMT const
- design.ts: add CANVAS constant; MONO already fixed to Roboto Mono in prior commit
- roles.ts: export shared MGMT set (single source of truth)
- index.html: add Roboto Mono to Google Fonts

Backend:
- mail.go: add 'mail' to RequirePages guard; fix listInboundMail info-disclosure (IS NULL join); fix markInboundRead ownership check
- auth.go: add 'mail' page to sales_officer, sales_head, bd_officer, bd_head roles
- approvals.go: initialise items as make([]item, 0) to prevent null JSON response
- fixed_deposit.go: change monetary amount fields from *float64 to *int64 (kobo rule)

+++ backend-go/handlers/fixed_deposit.go
@@ -85,11 +85,11 @@ func fdCreateTransaction(db *core.DB) http.HandlerFunc {
 			TransactionDate string   `json:"transaction_date"`
 			CustomerName    string   `json:"customer_name"`
 			TransactionType string   `json:"transaction_type"` // inflow | liquidation
-			Principal       *float64 `json:"principal"`
-			InterestPaid    *float64 `json:"interest_paid"`
-			GrossAmount     *float64 `json:"gross_amount"`
-			USDAmount       *float64 `json:"usd_amount"`
-			NGNAmount       *float64 `json:"ngn_amount"`
+			Principal       *int64   `json:"principal"`
+			InterestPaid    *int64   `json:"interest_paid"`
+			GrossAmount     *int64   `json:"gross_amount"`
+			USDAmount       *int64   `json:"usd_amount"`
+			NGNAmount       *int64   `json:"ngn_amount"`
 			Currency        string   `json:"currency"`
 			Location        string   `json:"location"`
 			AccountOfficer  string   `json:"account_officer"`
